Restructered code
