+++ code/setup/FolderDescription.md
@@ -1,11 +1,2 @@
 # Setup
 Contains all the building blocks of the project
-
-<ul>
-    <li><code>health_log.py</code></li>
-    Contains implementation of <code>Calender</code> class, which creates the calender you see in the Health Log aspect of the appplication
-    <li><code>setup_questions.py</code></li>
-    Contains <code>Questions</code> class, which is meant to be inherited from in order to add the functions for setting up the program. This was mainly created to avoid making <code>main.py</code> too large
-    <li><code>setup.py</code></li>
-    Defines the function <code>setup_logging</code> to ensure easy access to loggers and runlogs
-</ul>
