+++ code/utils/__init__.py
@@ -1,5 +1,4 @@
 from utils.parse_json import jsonUtils
 from utils.mcq import MCQbuiler
 from utils.data_classes import Question, CustomQuestion, UserInfo, InformationSheet, ActionButton
-from utils.special import Settings
-from utils.generic import set_theme, FileHandler, InformationPages
+from utils.generic import set_theme, FileHandler, UseLogger
