+++ json/preferences.json
@@ -1,4 +1,1 @@
-{
-    "appearance_theme": "dark",
-    "setup_finished": true
-}
+{}
