+++ json/user-data.json
@@ -1,4 +1,1 @@
-{
-    "birth_year": 1980,
-    "gender": "Male"
-}
+{}
